testing_requirements.txt
django-nose
mock